Removed obsolete files. Also renamed integration_new.py to integration.py. integration_new.py was the merging of the three old, removed integration tests.
