--- config_proxy302
@@ -1,10 +1,0 @@
-port 8082;
-
-path / ProxyHandler {
-    remote_host http://ucla.edu;
-    remote_port 80;
-}
-
-default NotFoundHandler {
-    root;
-}
